move global loader to its own file
