+++ components/GlobalLoader.tsx
@@ -1,0 +1,83 @@
+/*
+ * Copyright (c) 2022 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of
+ * the GNU Affero General Public License v3.0. You should have received a copy of the
+ * GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { CSSTransition } from 'react-transition-group';
+import { css } from '@icgc-argo/uikit';
+import DnaLoader from '@icgc-argo/uikit/DnaLoader';
+import Modal from '@icgc-argo/uikit/Modal';
+import { FadingDiv } from './Fader';
+import { fillAvailableHeight, fillAvailableWidth } from './Modal';
+
+export const loaderPortalRef = React.createRef<HTMLDivElement>();
+const GlobalLoaderView = ({ isGlobalLoading }: { isGlobalLoading: boolean }) => {
+  const ref = loaderPortalRef.current;
+  const fadeIn = 400;
+  const fadeOut = 600;
+  return ref
+    ? ReactDOM.createPortal(
+        <CSSTransition in={isGlobalLoading} timeout={fadeIn} classNames="on" unmountOnExit>
+          {() => (
+            <FadingDiv enterAnimationLength={fadeIn} exitAnimationLength={fadeOut}>
+              <Modal.Overlay
+                css={css`
+                  ${fillAvailableWidth}
+                  ${fillAvailableHeight}
+                  @media (min-width: 768px) {
+                    width: 100vw;
+                    height: 100vh;
+                  }
+                `}
+              >
+                <DnaLoader />
+              </Modal.Overlay>
+            </FadingDiv>
+          )}
+        </CSSTransition>,
+        ref,
+      )
+    : null;
+};
+
+const GlobalLoadingContext = React.createContext({
+  isGlobalLoading: false,
+  setGlobalLoading: (isGlobalLoading: boolean) => {},
+});
+
+const useGlobalLoader = () => React.useContext(GlobalLoadingContext);
+
+export const GlobalLoaderProvider = ({
+  children,
+  startWithGlobalLoader,
+}: {
+  children: React.ReactNode | React.ReactNodeArray;
+  startWithGlobalLoader: boolean;
+}) => {
+  const [isGlobalLoading, setGlobalLoading] = React.useState(startWithGlobalLoader || false);
+
+  return (
+    <GlobalLoadingContext.Provider value={{ isGlobalLoading, setGlobalLoading }}>
+      {children}
+      <GlobalLoaderView isGlobalLoading={isGlobalLoading} />
+    </GlobalLoadingContext.Provider>
+  );
+};
+
+export default useGlobalLoader;
